v2.1.1 (3/14 5:00 PM)
3/14 5:00 PM
The commit has the progress and tasks related to the development of the area mode functionality. The to-do list includes updating a corrected version, generating test point coordinates, developing area mode logic, generating commands for the test procedure, and figuring out how to perform each point test and store results in a variable. The document also includes a draft for the generate_gcode() method and a description of the area mode logic. Finally, the commit mentions some other topics related to Marlin and G-code.